docs: generate api docs (#1)
Adds DocC generated API docs as a CI job on main, published
[here](https://square.github.io/swift-modals/documentation/).

This job uses a [new feature of
DocC](https://forums.swift.org/t/a-preview-of-doccs-support-for-combined-documentation/74373)
to merge modules with a landing page. The inter-module symbol links
don't work yet; I think maybe there's just no way to enable that in
`xcodebuild docbuild`. But the joined navigation and landing page are a
huge improvement.